+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 * Allow few commands' execution without any Eask-file (#159)
 * Add command `cat` (#160)
 * Improve messages while processing package transaction (#164)
+* Improve messages while task cleaning all (#165)
 
 ## 0.8.x
 > Released Mar 08, 2023